@@ -4,7 +4,7 @@ description: A starting point for Dart libraries or applications.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   meta: ^1.3.0
@@ -13,4 +13,4 @@ dependencies:
 
 dev_dependencies:
   pedantic: ^1.11.0
-  test: ^1.17.0
+  test: ^1.17.1
