@@ -10,14 +10,25 @@ class User {
   final String name;
   final String displayName;
   final String email;
-  final String password;
+  final String? password;
   final group = IsarLink<UserGroup>();
 
-  User(this.id, {this.name, this.displayName, this.email, this.password});
+  User(this.id,
+      {required this.name, required this.displayName, required this.email, this.password});
 
-  User copyWith({String name, String displayName, String email, String password}) => User(id,
+  User copyWith({String? name, String? displayName, String? email, String? password}) => User(id,
       name: name ?? this.name,
       displayName: displayName ?? this.displayName,
       email: email ?? this.email,
       password: password ?? this.password);
+
+  User.fromJson(Map<String, dynamic> json)
+      : id = json['id'],
+        name = json['name'],
+        displayName = json['display-name'],
+        email = json['email'],
+        password = json['password'];
+
+  Map<String, dynamic> toJson() =>
+      {'id': id, 'name': name, 'display-name': displayName, 'email': email};
 }
